WIP, delete enpoint working, missing refetch
