+++ src/api/booksApi/baseQuery.ts
@@ -1,0 +1,8 @@
+import { fetchBaseQuery } from "@reduxjs/toolkit/dist/query";
+
+export const booksBaseQuery = fetchBaseQuery({
+  baseUrl: "http://localhost:3000/",
+  prepareHeaders: (headers) => {
+    headers.set("Content-Type", "application/json");
+  },
+});
